+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [Unreleased]
 
+### Fixed
+- Added translation for nutritient-value label placeholder
+  [#596](https://github.com/nextcloud/cookbook/pull/596) @seyfeb
 ## 0.8.1 - 2021-02-15
 
 ### Added